changed to new package io.github.astrapi69
