+++ gradle.properties
@@ -2,7 +2,7 @@
 # project properties #
 ######################
 projectVersion=5-SNAPSHOT
-groupPackage=de.alpharogroup
+groupPackage=io.github.astrapi69
 projectSourceCompatibility=1.8
 projectInceptionYear=2015
 projectHolderUsername=lightblueseas
