--- src/main/java/de/alpharogroup/package.html
@@ -1,11 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-<html>
-  <head>
-    <title>
-      de.alpharogroup
-    </title>
-  </head>
-  <body>
-    The main package from the Jaulp project.
-  </body>
-</html>
