+++ src/main/java/io/github/astrapi69/tree/LazyTreeElement.java
@@ -22,10 +22,8 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.tree;
+package io.github.astrapi69.tree;
 
-import de.alpharogroup.tree.api.ILinkedNode;
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -33,31 +31,24 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import lombok.experimental.FieldDefaults;
 
 /**
- * The class {@link LinkedNode} can have only one next element and one previous element
- *
- * @param <T>
- *            the generic type of the value
+ * The class {@link LazyTreeElement} represents a lazy tree element that does not know about
+ * the parent or children
  */
 @Getter
 @Setter
 @EqualsAndHashCode
-@ToString(exclude = { "next" })
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class LinkedNode<T> implements ILinkedNode<T>
+public class LazyTreeElement<T>
 {
 
-	/** The next. */
-	ILinkedNode<T> next;
-
-	/** The previous. */
-	ILinkedNode<T> previous;
-
 	/** The value. */
-	T value;
+	private T value;
+
+	/** The flag that indicates if this tree element is a leaf. */
+	private boolean leaf;
 }
