+++ src/main/java/io/github/astrapi69/tree/api/IChainableTreeNode.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.tree.api;
+package io.github.astrapi69.tree.api;
 
 import java.util.LinkedList;
 import java.util.List;
