+++ src/main/java/io/github/astrapi69/tree/api/package.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <title>
-		de.alpharogroup.tree.api
+		io.github.astrapi69.tree.api
     </title>
   </head>
   <body>
