+++ src/main/java/io/github/astrapi69/tree/package.html
@@ -2,10 +2,10 @@
 <html>
   <head>
     <title>
-		de.alpharogroup.tree
+      io.github.astrapi69.tree
     </title>
   </head>
   <body>
-    Provides classes for abstract trees and nodes.
+    Provides classes for abstract trees and nodes
   </body>
 </html>
