+++ src/test/resources/testng.xml
@@ -3,8 +3,8 @@
 <suite name="Suite">
   <test name="Test">
     <classes>
-      <class name="de.alpharogroup.tree.TreeElementTest"/>
-      <class name="de.alpharogroup.tree.TreeNodeTest"/>
+      <class name="io.github.astrapi69.tree.TreeElementTest"/>
+      <class name="io.github.astrapi69.tree.TreeNodeTest"/>
     </classes>
   </test> <!-- Test -->
 </suite> <!-- Suite -->
